remove all dynamic metric types
Removes all dynamic metric types.
